connect redux store directly to StudentSummary
