+++ modules/web/routes/create/method-upload.js
@@ -3,7 +3,7 @@ import DropZone from 'react-dropzone'
 import map from 'lodash/map'
 import Button from '../../components/button'
 import List from '../../components/list'
-import StudentSummary from '../../modules/student/student-summary'
+import { StudentSummary } from '../../modules/student/student-summary'
 import { initStudent } from '../../redux/students/actions/init-student'
 import withRouter from 'react-router/lib/withRouter'
 import { connect } from 'react-redux'
